Adds simple_imu_test + many bugfixes.
simple_imu_test is a test of the kalman filter on actual IMU values. It
approximates sin(x) as x since it's not an EKF -- everything has to be linear.
So this is only accurate for small angles.

This is also significant since it uses graphics code to visualize the data. This
is currently stored in a separate repository.