Pushing UNFINISHED convolution layer
Pushing this unfinished so I can work on it tomorrow from my laptop on caltrain.
There's no wifi on the train, so I need to push it from desktop and pull it from
laptop now, so that I'll have my unfinished work on my laptop on the train. I
know this is silly and I could have just sent myself a local patch, but it's
more convenient to just run git push and git pull.