Some work on getting backprop working.
This involved modifying layer_impl interface to move the work of
calculating gradients into each layer instead of doing it automatically.
This makes the process way way more efficient, to the point where it's
tractable (previously, generically and automatically calculating the
derivative of a convolutional network was causing OOM on my machine).