Optimized Training
Data no longer needs to be shuffled back and forth from GPU to CPU and back when
Training -- The forward pass computes all the layer inputs and then keeps them
in-place on the GPU for the backwards pass. This reduces memory consumption by a
factor of 2 and significantly speeds up training!