After many tears, fixed conv input kernels.
It was another underflow bug in the kernel source code.

To aid in debugging, I'm adding extras/example_kernel.cc which just
contains some example code that gets auto-generated by this framework.
Normally this stage is transparent because the generated code gets
passed straight to the opencl compiler, however I'm adding it to show
what goes on behind the scenes.